menu_favo: round bar Y position to single before truncation (89.24%)

+++ src/menu_favo.cpp
@@ -168,8 +168,8 @@ void CMenuPcs::FavoDraw()
 	FoodRank* rank = s_rank;
 	for (int i = 0; i < 8; i++) {
 		int barX = drawEntry->x + drawEntry->w + 0x18;
-		int barY = static_cast<int>((static_cast<float>(drawEntry->h) - 24.0f) * 0.5 +
-		                            static_cast<float>(drawEntry->y));
+		int barY = static_cast<int>(static_cast<float>((static_cast<float>(drawEntry->h) - 24.0f) * 0.5 +
+		                            static_cast<float>(drawEntry->y)));
 		DrawSingBar(barX, barY, rank->score, drawEntry->alpha);
 		rank++;
 		drawEntry++;
